moving Equivalency interface out of internal package since it is part of the public API
